+++ NEWS.md
@@ -1,3 +1,5 @@
+# baguette (development version)
+
 # baguette 0.2.0
 
 * Model definitions were moved to parsnip (#47)